Updated settings page
