+++ core/api/src/actions/navigation.ts
@@ -115,7 +115,7 @@ export class NavigationList extends OptionallyAuthenticatedAction {
         platformItems.push({
           type: "link",
           title: "Settings",
-          href: "/settings",
+          href: "/settings/core",
         });
 
         platformItems.push({
